+++ R/plotting.R
@@ -179,18 +179,18 @@ plot_cell_trajectory <- function(
     g <- ggplot(data = data_df, aes(x = data_dim_1, y = data_dim_2))
   }
   if (show_tree) {
-    g <- g + geom_segment(aes_string(x = "source_prin_graph_dim_1", y = "source_prin_graph_dim_2", xend = "target_prin_graph_dim_1", yend = "target_prin_graph_dim_2"), size = cell_link_size, linetype = "solid", na.rm = TRUE, data = edge_df)
+    g <- g + geom_segment(aes(x = source_prin_graph_dim_1, y = source_prin_graph_dim_2, xend = target_prin_graph_dim_1, yend = target_prin_graph_dim_2), linewidth = cell_link_size, linetype = "solid", na.rm = TRUE, data = edge_df)
   }
 
   if (is.null(markers_exprs) == FALSE && nrow(markers_exprs) > 0) {
     if (use_color_gradient) {
     } else {
-      g <- g + geom_point(aes_string(color = color_by), na.rm = TRUE)
+      g <- g + geom_point(aes(color = !!rlang::sym(color_by)), na.rm = TRUE)
     }
   } else {
     if (use_color_gradient) {
     } else {
-      g <- g + geom_point(aes_string(color = color_by), size = I(cell_size), na.rm = TRUE)
+      g <- g + geom_point(aes(color = !!rlang::sym(color_by)), size = I(cell_size), na.rm = TRUE)
     }
   }
 
@@ -201,10 +201,10 @@ plot_cell_trajectory <- function(
       dplyr::mutate(branch_point_idx = seq_len(dplyr::n()))
 
     g <- g +
-      geom_point(aes_string(x = "prin_graph_dim_1", y = "prin_graph_dim_2"),
+      geom_point(aes(x = prin_graph_dim_1, y = prin_graph_dim_2),
         size = 5, na.rm = TRUE, branch_point_df
       ) +
-      geom_text(aes_string(x = "prin_graph_dim_1", y = "prin_graph_dim_2", label = "branch_point_idx"),
+      geom_text(aes(x = prin_graph_dim_1, y = prin_graph_dim_2, label = branch_point_idx),
         size = 4, color = "white", na.rm = TRUE, branch_point_df
       )
   }
@@ -369,9 +369,9 @@ plot_genes_violin <- function(
       levels = panel_order
     )
   }
-  q <- ggplot(aes_string(x = grouping, y = "expression"), data = cds_exprs)
+  q <- ggplot(aes(x = !!rlang::sym(grouping), y = expression), data = cds_exprs)
   if (is.null(color_by) == FALSE) {
-    q <- q + geom_violin(aes_string(fill = color_by))
+    q <- q + geom_violin(aes(fill = !!rlang::sym(color_by)))
   } else {
     q <- q + geom_violin()
   }
@@ -380,14 +380,18 @@ plot_genes_violin <- function(
       fun.data = "mean_cl_boot",
       size = 0.2
     )
-    q <- q + stat_summary(
-      aes_string(
-        x = grouping, y = "expression",
-        group = color_by
-      ),
-      fun.data = "mean_cl_boot",
-      size = 0.2, geom = "line"
-    )
+    if (is.null(color_by) == FALSE) {
+      q <- q + stat_summary(
+        aes(group = !!rlang::sym(color_by)),
+        fun.data = "mean_cl_boot",
+        linewidth = 0.2, geom = "line"
+      )
+    } else {
+      q <- q + stat_summary(
+        fun.data = "mean_cl_boot",
+        linewidth = 0.2, geom = "line"
+      )
+    }
   }
   q <- q + facet_wrap(~feature_label,
     nrow = nrow,
@@ -493,16 +497,21 @@ plot_genes_jitter <- function(
     cds_exprs$feature_label <- factor(cds_exprs$feature_label, levels = panel_order)
   }
 
-  q <- ggplot(aes_string(x = grouping, y = "expression"), data = cds_exprs)
+  q <- ggplot(aes(x = !!rlang::sym(grouping), y = expression), data = cds_exprs)
 
   if (is.null(color_by) == FALSE) {
-    q <- q + geom_jitter(aes_string(color = color_by), size = I(cell_size))
+    q <- q + geom_jitter(aes(color = !!rlang::sym(color_by)), size = I(cell_size))
   } else {
     q <- q + geom_jitter(size = I(cell_size))
   }
   if (plot_trend == TRUE) {
-    q <- q + stat_summary(aes_string(color = color_by), fun.data = "mean_cl_boot", size = 0.35)
-    q <- q + stat_summary(aes_string(x = grouping, y = "expression", color = color_by, group = color_by), fun.data = "mean_cl_boot", size = 0.35, geom = "line")
+    if (is.null(color_by) == FALSE) {
+      q <- q + stat_summary(aes(color = !!rlang::sym(color_by)), fun.data = "mean_cl_boot", size = 0.35)
+      q <- q + stat_summary(aes(color = !!rlang::sym(color_by), group = !!rlang::sym(color_by)), fun.data = "mean_cl_boot", linewidth = 0.35, geom = "line")
+    } else {
+      q <- q + stat_summary(fun.data = "mean_cl_boot", size = 0.35)
+      q <- q + stat_summary(fun.data = "mean_cl_boot", linewidth = 0.35, geom = "line")
+    }
   }
 
   q <- q + scale_y_log10() + facet_wrap(~feature_label, nrow = nrow, ncol = ncol, scales = "free_y")
@@ -610,13 +619,13 @@ plot_genes_positive_cells <- function(
 
   if (plot_as_fraction) {
     marker_counts$target_fraction <- marker_counts$target_fraction * 100
-    qp <- ggplot(aes_string(x = grouping, y = "target_fraction", fill = grouping), data = marker_counts) +
+    qp <- ggplot(aes(x = !!rlang::sym(grouping), y = target_fraction, fill = !!rlang::sym(grouping)), data = marker_counts) +
       ylab("Cells (percent)")
     if (is.null(plot_limits) == FALSE) {
       qp <- qp + scale_y_continuous(limits = plot_limits)
     }
   } else {
-    qp <- ggplot(aes_string(x = grouping, y = "target", fill = grouping), data = marker_counts) +
+    qp <- ggplot(aes(x = !!rlang::sym(grouping), y = target, fill = !!rlang::sym(grouping)), data = marker_counts) +
       ylab("Cells")
   }
 
@@ -735,7 +744,7 @@ plot_genes_in_pseudotime <- function(
   }
   q <- ggplot(aes(Pseudotime, expression), data = cds_exprs)
   if (is.null(color_by) == FALSE) {
-    q <- q + geom_point(aes_string(color = color_by), size = I(cell_size), position = position_jitter(horizontal_jitter, vertical_jitter))
+    q <- q + geom_point(aes(color = !!rlang::sym(color_by)), size = I(cell_size), position = position_jitter(horizontal_jitter, vertical_jitter))
   } else {
     q <- q + geom_point(size = I(cell_size), position = position_jitter(horizontal_jitter, vertical_jitter))
   }
@@ -1190,7 +1199,7 @@ plot_genes_branched_pseudotime <- function(
 
   q <- ggplot(aes(Pseudotime, expression), data = cds_exprs)
   if (is.null(color_by) == FALSE) {
-    q <- q + geom_point(aes_string(color = color_by), size = I(cell_size))
+    q <- q + geom_point(aes(color = !!rlang::sym(color_by)), size = I(cell_size))
   }
   if (is.null(reducedModelFormulaStr) == FALSE) {
     q <- q + scale_y_log10() + facet_wrap(~ feature_label +
@@ -1205,14 +1214,11 @@ plot_genes_branched_pseudotime <- function(
       method = "loess"
     )
   } else if (method == "fitting") {
-    q <- q + geom_line(aes_string(
-      x = "Pseudotime", y = "full_model_expectation",
-      linetype = "Branch"
-    ), data = cds_exprs)
+    q <- q + geom_line(aes(x = Pseudotime, y = full_model_expectation, linetype = Branch), data = cds_exprs)
   }
 
   if (!is.null(reducedModelFormulaStr)) {
-    q <- q + geom_line(aes_string(x = "Pseudotime", y = "reduced_model_expectation"),
+    q <- q + geom_line(aes(x = Pseudotime, y = reduced_model_expectation),
       color = "black", linetype = 2, data = cds_exprs
     )
   }
@@ -1725,7 +1731,7 @@ plot_cell_clusters <- function(
     g <- g + geom_point(aes(color = log10(value + 0.1)), size = I(cell_size), na.rm = TRUE) +
       viridis::scale_color_viridis(name = paste0("log10(value + 0.1)"), ...)
   } else {
-    g <- g + geom_point(aes_string(color = color_by), size = I(cell_size), na.rm = TRUE)
+    g <- g + geom_point(aes(color = !!rlang::sym(color_by)), size = I(cell_size), na.rm = TRUE)
   }
 
   g <- g +
@@ -1989,7 +1995,7 @@ plot_complex_cell_trajectory <- function(
     stop("You must first call orderCells() before using this function")
   }
 
-  edge_list <- as.data.frame(igraph::get.edgelist(dp_mst))
+  edge_list <- as.data.frame(igraph::as_edgelist(dp_mst))
   colnames(edge_list) <- c("source", "target")
 
   edge_df <- merge(ica_space_df, edge_list, by.x = "sample_name", by.y = "source", all = TRUE)
@@ -2029,22 +2035,22 @@ plot_complex_cell_trajectory <- function(
     g <- ggplot(data = data_df, aes(x = data_dim_1, y = data_dim_2))
   }
   if (show_tree) {
-    g <- g + geom_segment(aes_string(x = "source_prin_graph_dim_1", y = "source_prin_graph_dim_2", xend = "target_prin_graph_dim_1", yend = "target_prin_graph_dim_2"), size = cell_link_size, linetype = "solid", na.rm = TRUE, data = edge_df)
+    g <- g + geom_segment(aes(x = source_prin_graph_dim_1, y = source_prin_graph_dim_2, xend = target_prin_graph_dim_1, yend = target_prin_graph_dim_2), linewidth = cell_link_size, linetype = "solid", na.rm = TRUE, data = edge_df)
   }
 
   if (is.null(markers_exprs) == FALSE && nrow(markers_exprs) > 0) {
     if (inherits(data_df[, color_by], "numeric")) {
-      g <- g + geom_jitter(aes_string(color = paste0("log10(", color_by, " + 0.1)")), size = I(cell_size), na.rm = TRUE, height = 5) +
+      g <- g + geom_jitter(aes(color = log10(!!rlang::sym(color_by) + 0.1)), size = I(cell_size), na.rm = TRUE, height = 5) +
         viridis::scale_color_viridis(name = paste0("log10(", color_by, ")"), ...)
     } else {
-      g <- g + geom_jitter(aes_string(color = color_by), size = I(cell_size), na.rm = TRUE, height = 5)
+      g <- g + geom_jitter(aes(color = !!rlang::sym(color_by)), size = I(cell_size), na.rm = TRUE, height = 5)
     }
   } else {
     if (inherits(data_df[, color_by], "numeric")) {
-      g <- g + geom_jitter(aes_string(color = paste0("log10(", color_by, " + 0.1)")), size = I(cell_size), na.rm = TRUE, height = 5) +
+      g <- g + geom_jitter(aes(color = log10(!!rlang::sym(color_by) + 0.1)), size = I(cell_size), na.rm = TRUE, height = 5) +
         viridis::scale_color_viridis(name = paste0("log10(", color_by, " + 0.1)"), ...)
     } else {
-      g <- g + geom_jitter(aes_string(color = color_by), size = I(cell_size), na.rm = TRUE, height = 5)
+      g <- g + geom_jitter(aes(color = !!rlang::sym(color_by)), size = I(cell_size), na.rm = TRUE, height = 5)
     }
   }
 
@@ -2054,10 +2060,10 @@ plot_complex_cell_trajectory <- function(
     branch_point_df$branch_point_idx <- match(branch_point_df$sample_name, mst_branch_nodes)
     branch_point_df <- branch_point_df[!duplicated(branch_point_df$branch_point_idx), ]
 
-    g <- g + geom_point(aes_string(x = "source_prin_graph_dim_1", y = "source_prin_graph_dim_2"),
+    g <- g + geom_point(aes(x = source_prin_graph_dim_1, y = source_prin_graph_dim_2),
       size = 2 * cell_size, na.rm = TRUE, data = branch_point_df
     ) +
-      geom_text(aes_string(x = "source_prin_graph_dim_1", y = "source_prin_graph_dim_2", label = "branch_point_idx"),
+      geom_text(aes(x = source_prin_graph_dim_1, y = source_prin_graph_dim_2, label = branch_point_idx),
         size = 1.5 * cell_size, color = "white", na.rm = TRUE, data = branch_point_df
       )
   }
@@ -2434,7 +2440,7 @@ plot_multiple_branches_pseudotime <- function(
 
   q <- ggplot(aes(Pseudotime, expression), data = cds_exprs)
   if (is.null(color_by) == FALSE) {
-    q <- q + geom_line(aes_string(color = color_by), size = I(cell_size))
+    q <- q + geom_line(aes(color = !!rlang::sym(color_by)), linewidth = I(cell_size))
   }
   q <- q + facet_wrap(~feature_label, nrow = nrow, ncol = ncol, scales = "free_y")