+++ tests/testthat/test.clusterCells.R
@@ -49,7 +49,7 @@ melted_dens_df <- reshape2::melt(Matrix::t(scale(Matrix::t(L))))
 # Plot the distribution of the standardized gene expression values.
 ggplot(melted_dens_df, aes(x = value)) +
   geom_density() +
-  stat_function(fun = dnorm, size = 0.5, color = "red") +
+  stat_function(fun = dnorm, linewidth = 0.5, color = "red") +
   xlab("Standardized log(FPKM)") +
   ylab("Density")